jv - renamed fixtures
