+++ frontend/src/stories/components/MenuItemReviews/MenuItemReviewsForm.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import MenuItemReviewsForm from "main/components/MenuItemReviews/MenuItemReviewsForm";
-import { menuItemReviewsFixtures } from 'fixtures/menuItemReviewsFixtures';
+import { menuItemReviewsFixtures } from 'fixtures/menuItemReviewsFixtures2';
 
 export default {
     title: 'components/MenuItemReviews/MenuItemReviewsForm',
