+++ frontend/src/tests/components/MenuItemReviews/MenuItemReviewsTable.test.js
@@ -1,5 +1,5 @@
 import { fireEvent, render, waitFor } from "@testing-library/react";
-import { menuItemReviewsFixtures, MenuItemReviewsFixtures } from "fixtures/menuItemReviewsFixtures";
+import { menuItemReviewsFixtures, MenuItemReviewsFixtures } from "fixtures/menuItemReviewsFixtures2";
 import MenuItemReviewsTable from "main/components/MenuItemReviews/MenuItemReviewsTable"
 import { QueryClient, QueryClientProvider } from "react-query";
 import { MemoryRouter } from "react-router-dom";
